Fix League Gothic font's missing diacritical marks, needed for French.
Displaying a title featuring the custom font, such as "Taggée", formatted
the "é" using the browser's font, while the rest of the text used the
desired custom font. Hideous.

The solution was to download the font's original OTF source file and
generate the fontface using fontsquirrel.com, which included all the glyphs.