Fix issue #63: vagrant fails with mmap errors during apt-get.
The Vagrantfile described a cache to store downloaded packages on the
local machine and exported this to the guest, to save it from
re-downloading files if the guest was rest.

Unfortunately, `apt-get` only works if its files are on a local
filesystem or in NFS mounted directories, but not the default VirtualBox
file sharing. Therefore, this would fail for anyone not using NFS.

The solution is to only enable the package cache if the user has enabled
NFS file sharing mode.